refactor: remove video app
