--- now/app/image_text_retrieval/app.py
@@ -1,66 +1,0 @@
-from typing import Dict, List, Tuple, Union
-
-from docarray import DocumentArray
-
-from now.app.base.app import JinaNOWApp
-from now.common.utils import _get_clip_apps_with_dict, common_setup, get_indexer_config
-from now.constants import CLIP_USES, Apps, Modalities
-from now.demo_data import DemoDatasetNames
-from now.now_dataclasses import UserInput
-
-
-class ImageTextRetrieval(JinaNOWApp):
-    def __init__(self):
-        super().__init__()
-
-    @property
-    def app_name(self) -> str:
-        return Apps.IMAGE_TEXT_RETRIEVAL
-
-    @property
-    def is_enabled(self) -> bool:
-        return True
-
-    @property
-    def description(self) -> str:
-        return 'Image-text search app'
-
-    @property
-    def input_modality(self) -> Union[Modalities, List[Modalities]]:
-        return [Modalities.IMAGE, Modalities.TEXT]
-
-    @property
-    def output_modality(self) -> Union[Modalities, List[Modalities]]:
-        return [Modalities.IMAGE, Modalities.TEXT]
-
-    @property
-    def required_docker_memory_in_gb(self) -> int:
-        return 8
-
-    @property
-    def finetune_datasets(self) -> [Tuple]:
-        return DemoDatasetNames.DEEP_FASHION, DemoDatasetNames.BIRD_SPECIES
-
-    def setup(
-        self, dataset: DocumentArray, user_input: UserInput, kubectl_path
-    ) -> Dict:
-        indexer_config = get_indexer_config(len(dataset))
-        encoder_with = _get_clip_apps_with_dict(user_input)
-        env_dict = common_setup(
-            app_instance=self,
-            user_input=user_input,
-            dataset=dataset,
-            encoder_uses=CLIP_USES[user_input.deployment_type][0],
-            encoder_with=encoder_with,
-            encoder_uses_with={
-                'pretrained_model_name_or_path': CLIP_USES[user_input.deployment_type][
-                    1
-                ]
-            },
-            pre_trained_embedding_size=CLIP_USES[user_input.deployment_type][2],
-            indexer_uses=indexer_config['indexer_uses'],
-            kubectl_path=kubectl_path,
-            indexer_resources=indexer_config['indexer_resources'],
-        )
-        super().setup(dataset=dataset, user_input=user_input, kubectl_path=kubectl_path)
-        return env_dict
