+++ tests/integration/test_transform_docarray.py
@@ -6,7 +6,7 @@ from docarray.typing import Image, Text
 from jina import Flow
 
 from now.app.base.transform_docarray import transform_uni_modal_data
-from now.app.image_text_retrieval.app import ImageTextRetrieval
+from now.app.search_app.app import ImageTextRetrieval
 from now.app.text_to_video.app import TextToVideo
 from now.constants import ACCESS_PATHS, EXTERNAL_CLIP_HOST, DatasetTypes
 from now.data_loading.data_loading import load_data
