+++ tests/unit/apps/test_preprocessing.py
@@ -5,7 +5,7 @@ from docarray import Document, DocumentArray
 
 from now.app.base.preprocess import preprocess_text
 from now.app.base.transform_docarray import transform_docarray
-from now.app.image_text_retrieval.app import ImageTextRetrieval
+from now.app.search_app.app import ImageTextRetrieval
 from now.app.text_to_video.app import TextToVideo
