+++ tests/unit/data_loading/test_data_loading.py
@@ -6,7 +6,7 @@ import pytest
 from docarray import Document, DocumentArray
 from pytest_mock import MockerFixture
 
-from now.app.image_text_retrieval.app import ImageTextRetrieval
+from now.app.search_app.app import ImageTextRetrieval
 from now.constants import DatasetTypes
 from now.data_loading.data_loading import _load_tags_from_json_if_needed, load_data
 from now.demo_data import DemoDatasetNames
